REPL log panel now uses the LineStyleListener approach for styling.
The StyledText widget offers two mutually exclusive ways for handling
line styling:
- style range methods, which store styles in the widget,
- LineStyleListener(s) that return styles whenever required.
Previously the first method was used, preventing the use of
LineStyleListeners unless one wants to override CCW styling.
By switching to the LineStyleListner approach, it's now possible for
other CCW components and plugins to participate in styling by adding new
listeners.
This change requires the caching of line styles because there is no
point in recomputing the style for text that does not change.