feat(repl): now you can choose the port of the embedded nRepl with -Dccw.nrepl.port
By refactoring startREPLServer and moving everything in clj sources, one property has been added in
order to select the port of the embedded nRepl: ccw.nrepl.port. Note that this is only effective
when ccw.nrepl.autostart is present (the name has been changed for coerence). The usage is simple,
just add -Dccw.nrepl.autostart -Dccw.nrepl.port=[number 1-65535]. If the port number is not in range
or absent, the first free randomly-choosen port will be used. The .classpath has been brought to
1.7.

BREAKING: Bump of tools.nrepl to 0.2.5, of Exec. Environment to 1.7; Old ccw.autostartrepl property
renamed in ccw.nrepl.autostart.