feat(repl): now you can enable the cider (default) middleware with -Dccw.nrepl.cider.enable
This feature allows to have the cider middleware and therefore full cider support on the embedded
Counterclockwise nRepl. Without it, a warning is raised and "apropos classpath complete eldoc..."
won't be working. The property ccw.nrepl.cider.enable needs to be present at runtime.

The new functionality is enabled by adding -Dccw.nrepl.cider.enable to the command line or in
Counterclockwise.ini.