feat(ccw.core): add pluggable ITextHover* through ccw.core.cljEditorTextHovers and hover pref page
This feature adds the ability of contributing and configuring ITextHover* instances to
Counterclockwise.
The extension point is ccw.core.cljEditorTextHovers whereas the schema (a cousin of Jdt's) is
contained in schema/cljEditorTextHovers.exsd. The big difference is that a ITextHover instance can
be returned by Clojure functions leveraging ccw.util.GenericExecutableExtension (see plugin.xml for
an example).
This patch also adds a reactive preference page where the user can enable, disable or change the
hover modifier. It has been rewritten from scratch using the Eclipse's databinding framework,
solving a couple of annoying bugs of the Jdt one. The page follows the MVVM pattern and needs
injection of the HoverModel. In order to achieve that, a custom IEclipseContext has been created and
stored in CCWPlugin together with init/clean functions.

The hover framework traces using "support/hover".

Conflicts:
	ccw.core/META-INF/MANIFEST.MF
	ccw.core/build.properties
	ccw.core/plugin.xml
	ccw.core/src/clj/ccw/editors/clojure/clojure_text_hover.clj
	ccw.core/src/java/ccw/CCWPlugin.java