feat(ccw.core-hovers): retarget text.showInformation, add hover base classes and css
The patch enhances the hover framework by adding a common set of base classes and a common style
that other future hovers (meta, source, etc...) can leverage. First and foremost, the IClojureHover
protocol has been introduced (see docstring_hover.clj for the usage). Second,
ccw.core/assets/hover-control-style.css is prepended to the displayed html in order to control the
output of HoverControlCreator and HoverEnrichedControlCreator.
Together with the above, the org.eclipse.ui.edit.text.showInformation action has been retargeted and
bound to IInformationPresenter/Provider which at the moment handles just one document partition and
delegates to ITextHover instances the information retrieval.

The refactoring of doc_utils.clj changes the way "no information" is handled. Nil is returned
instead of "no doc found" so that clients can decide how to display that condition. In particular,
the current implementation displays a message in the status bar.