feat(ccw.core.scanners): ClojureTokenScanner reacts to changes (syntax coloring in pref and editor)
Completely revisited the management of ClojureTokenScanner and everything related to syntax
coloring: the algorithm stays the same of course but now the scanner reacts to preferences changes
and is able to adapt its syntax coloring. This is particularly evident in
SyntaxColoringPreferencePage, which has been refactored a bit in order to keep the previewer always
visible at the bottom of the key list.

ClojureSourceViewer, ClojureSourceViewerConfiguration and ClojureTokenScanner have become listeners
to coloring preference changes, they respect configure/unconfigure life-cycle.

This patch also refactors the location of scanner classes (ccw.editors.clojure.scanners).