fix(ccw.core-colors): Fix "Color Disposed" Exception in ClojureSourceViewer.EditorColors
Given that ColorRegistry handles the disposal of the colors itself, and that all colors should be
cached (because gotten through CCWPlugin.getColor), then there is no need for manual .dispose()
calls in ClojureSourceViewer.initializeViewerColors.  At the same time, no other color caches or
managers are needed, like the one that was in SyntaxColoringPreferencePage, and ClojureColorManager
could be removed (this one left to the project maintainer).