fix(ccw.core-scanners): Fix for black parens following the very first form (correctly displayed)
In an attempt to increase the parsing speed in the ClojureTokenScanner, an -else if- had been
introduced in ClojureTokenScanner.nextToken() in order to skip the whitespace processing. This was
causing all the first parens of all the forms but the first to be displayed in black.

The improvement has been commented out, waiting for some better patch.