feat(ccw.core): Create ccw/core.clj and refactor delayed atom paradigm to it
The patch creates the ccw.core namespace + clj for including utils that we cannot classify in
bundle.clj, eclipse.clj, ...
It also refactors the delayed atom pattern (copyright Laurant Petit) into it with tests.

The pattern is applied to ccw.core.launch were it was originally introduced. This breaks some
private defn name.