fix(tracing): Re-enables tracing from ccw.core.trace
The patch explores and fixes the reason why tracing from clj files was missing. On the Java side the
specification of tracing option with option path preceded by a forward slash worked but in
ccw.core.trace, or better ccw.trace macros, the forward slash was missing. Moreover, evident
warnings have been added as documentation in ITracer.

Note that it is absolutely necessary to specify the tracing option this way, as deep down
EclipseDebugTrace, a simple final String option = bundleSymbolicName + optionPath is performed.