fix(ccw.core-hovers): Bug with initialization of the hovers, improved consistency
The atom part of hover_support.clj has been reworked as it was "code smelling" (as Laurent rightly
put it). The state has been encapsulated in an single atom and it is read/written all
together.
Moreover the update WritableList was throwing because the code was not executed in the right
Realm (needs to be the one from SWTObservables, which needs a Display itself). Fixed as well.