fix(ccw.core-hovers): Checked state lost on extension update in HoverPreferencePage
The problem was happening in the remote case that at runtime a hover extension point was
changed. The checked flag in the HoverPreferencePage was lost and all the items where left
unchecked.  To reproduce it, I telnetted to the osgi console and uninstalled a dummy bundle which
was defining its own hovers.
The fix disables automatic updating and changes a little bit the "when" the update from
HoverViewModel.checkedSet model to CheckedTableViewer target happens. The Eclipse databinding
framework is still immature, especially in the documentation.