fix(ccw.core-hovers): hover computed size mismatch due to <pre> and cut docstring
All the <pre> tags in the hover's html were causing the docstring hovers to be cut, both in the
horizontal and vertical sense. After investigating I noticed that BrowserInformationControl
calculates its width in a way that does not consider that <pre> tags are monospace, therefore
bigger.

Two options arised:
1) Change the <pre> tag in something different, losing the monospace font.
2) Find a way to consider the bigger font size of <pre>.

This solution hacks together option number 2.