fix(ccw.core-hovers): no hover was shown in Repl's input area
Little refactoring of getCorrespondingREPL and getSafeToolingConnection that from ClojureEditor are
now part of the IReplProvider interface. Note that the implementation has been duplicated in both
ClojureEditor and REPLView as temporary patch to the problem.  The addition of IReplProvider is just
the first step of a bigger refactoring involving how to handle Repl sessions in their entirety, that
will be part of another pull request.