refactor(asyncExec): all asyncExec/syncExec is now in DisplayUtil and everything uses it
In order to have a single point of failure, all the calls to display.asyncExec and display.syncExec
are now pointing to the DisplayUtil class.  This required some refactoring of both .java and .clj
files: in particular some code that originally was in eclipse.clj now has been moved to
swt.clj (there was a TODO about it).