feat(ccw.core.test): add Test and Generator... menu and fixtures for it
This patch enables the user to record tests using the SWTBot generator and therefore ease the pain
of bug reporting. It is also useful for debugging purposes and/or test enhancing.
A contributed menu entry called "Generator..." in the Test menu will open the SWTBot interface. In
the process, the MANIFEST.MF has been updated with the new version of SWTBot and updates it with the
dependencies taken from the SWTBot samples. Hamcrest imports were also needed. Tests were added.

Note1: the Clojure menu is always shown now because the user might want to call Test Generator even
when the active part is not the ClojureEditor.
Note2: the patch also performs a little refactoring of the command contributions.