feat(leiningen): isolate update dependencies execution and added cancelability
The patch introduces two features. The first disallow the possibility to run concurrent dependency
updates. An atom has been added in ccw.leiningen.classpath-container in order to control this.
The second is WorkspaceJob cancel support: the implementation can be polished with macros but at the
moment functions are passed around (especially in get-project-dependencies) in order to throw
OperationCancelledException in the appropriate place and a new cancelable-workspace-job has been
added in ccw.eclipse.

The cancel response is not immediate as only get-project-dependencies has been harnessed with
breaking capabilities.