fix(folding): fix for #799 and better projection management
This patch solves the dreaded issue #799 by adding better projection management on
IClojureEditor. It introduces an explicit enableProjection on IClojureEditor and a new preference:
PreferenceConstants/EDITOR_FOLDING_PROJECTION_ENABLED.  The trick for issue #799 was to add a
listener that triggers every time AbstractTextEditor/PREFERENCE_COLOR_BACKGROUND changes and turns
off and on the projection fixture on all the editors. By doing this, they paint their
projection/folding vertical column with the right color.